requirements.txt
anyio==3.6.2
asgiref==3.5.2
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.70.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
passlib==1.7.4
pydantic==1.10.2
PyJWT==1.7.1
python-decouple==3.3
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.16.0
typing-extensions==4.4.0
uvicorn==0.15.0
bcrypt==4.0.1